MovieDetails moved outside the MovieList folder
