+++ src/core/App/App.js
@@ -2,7 +2,7 @@ import { Redirect, Route, Switch } from "react-router-dom";
 import { Navigation } from "../../common/Navigation";
 import { MoviesList } from "../../features/MoviesList";
 import { PeopleList } from "../../features/PeopleList";
-import { MovieDetails } from "../../features/MoviesList/MovieDetails"
+import { MovieDetails } from "../../features/MovieDetails"
 import { toMovie, toMovies, toPeople, toPerson } from "./routes";
