+++ src/core/rootSaga.js
@@ -2,7 +2,7 @@ import { all } from "redux-saga/effects";
 import { popularMoviesSaga } from "../features/MoviesList/popularMoviesSaga";
 import { genresSaga } from "../features/MoviesList/Genres/genresSaga";
 import { popularPeopleSaga} from "../features/PeopleList/popularPeopleSaga";
-import { movieDetailsSaga } from "../features/MoviesList/MovieDetails/movieDetailsSaga";
+import { movieDetailsSaga } from "../features/MovieDetails/movieDetailsSaga";
 export default function* rootSaga() {
   yield all([
     popularMoviesSaga(),
