+++ src/features/MoviesList/index.js
@@ -1,7 +1,7 @@
 import { useDispatch, useSelector } from "react-redux";
 import { Container } from "../../common/Container/styled";
 import { MainHeader } from "../../common/MainHeader";
-import { List } from "./styled";
+import { List, StyledLink } from "./styled";
 import { useEffect } from "react";
 import {
   fetchPopularMovies,
@@ -10,7 +10,7 @@ import {
 } from "./popularMoviesSlice";
 import { MovieTile } from "../../common/MovieTile";
 import { ErrorPage } from "../../common/ErrorPage";
-
+import { toMovie } from "../../core/App/routes";
 import { Pagination } from "../../common/Pagination";
 
 export const MoviesList = () => {
@@ -33,11 +33,13 @@ export const MoviesList = () => {
         <List>
           {fetchResult.results.map((movie) => (
             <div key={movie.id}>
+              <StyledLink to={toMovie({movieId: movie.id})} >
               <MovieTile 
               movie={movie} 
               id={movie.id}
               genres={movie.genre_ids}
               />
+              </StyledLink>
             </div>
           ))}
         </List>
